feat: Add DataReferenc.Json<T>
