+++ IntroToCSharp/Shared/Layout/MainLayout.razor
@@ -106,7 +106,7 @@
         {
             if (user == null) return;
             if (user.DarkMode != null)
-                user.DarkMode.DataChanged += HandleDarkMode;
+                user.DarkMode.DataChangedEvent += HandleDarkMode;
         });
     }
